Switch to UTF8String.

+++ TextEditor.Win32/TextEditor.Win32/MainWindow.hpp
@@ -8,7 +8,7 @@
 class MainWindow {
 private:
     static HWND hwnd;
-    static UINT width, height;
+    static WORD width, height;
     static TextBox textBox;
 
     static LRESULT CALLBACK Callback(const HWND hwnd, const UINT uMsg, const WPARAM wParam, const LPARAM lParam);
